aumento risorse base implemented

+++ src/main/java/com/giancotsu/owar/projection/SviluppoCostiProjection.java
@@ -3,7 +3,7 @@ package com.giancotsu.owar.projection;
 public interface SviluppoCostiProjection {
 
     Integer getLivello();
-    Long getMoltiplicatore();
-    Long getCosto();
+    Double getMoltiplicatore();
+    Double getCosto();
     String getRisorsa();
 }
